+++ src/screens/Home/Views/Setting/Vertical/Main.tsx
@@ -11,7 +11,6 @@ import Backup from '../settings/Backup'
 import Other from '../settings/Other'
 import Version from '../settings/Version'
 import About from '../settings/About'
-import CustomSource from '../settings/CustomSource'
 import { createStyle } from '@/utils/tools'
 import { SETTING_SCREENS, type SettingScreenIds } from '../Main'
 
@@ -41,7 +40,6 @@ const ListItem = memo(({
     case 'other': return <Other />
     case 'version': return <Version />
     case 'about': return <About />
-    case 'custom_source': return <CustomSource />
     case 'basic': return <Basic />
   }
 }, () => true)